+++ greentest/test__monkey_sigchld.py
@@ -8,7 +8,7 @@ import gevent.monkey
 gevent.monkey.patch_all()
 
 pid = None
-awaiting_child = [True]
+awaiting_child = []
 
 
 def handle_sigchld(*args):
@@ -26,24 +26,44 @@ if hasattr(signal, 'SIGCHLD'):
     handler = signal.getsignal(signal.SIGCHLD)
     assert signal.getsignal(signal.SIGCHLD) is handle_sigchld, handler
 
-    pid = os.fork()
-    if not pid:
-        # child
-        gevent.sleep(0.2)
-        sys.exit(0)
+    if hasattr(os, 'forkpty'):
+        def forkpty():
+            # For printing in errors
+            return os.forkpty()[0]
+        funcs = (os.fork, forkpty)
     else:
-        with gevent.Timeout(1):
-            while awaiting_child:
-                gevent.sleep(0.01)
-            # We should now be able to waitpid() for an arbitrary child
-            wpid, status = os.waitpid(-1, os.WNOHANG)
-            assert wpid == pid
-            # And a second call should raise ECHILD
-            try:
-                wpid, status = os.waitpid(-1, os.WNOHANG)
-                raise AssertionError("Should not be able to wait again")
-            except OSError as e:
-                assert e.errno == errno.ECHILD
+        funcs = (os.fork,)
+
+    for func in funcs:
+        awaiting_child = [True]
+        pid = func()
+        if not pid:
+            # child
+            gevent.sleep(0.3)
             sys.exit(0)
+        else:
+            timeout = gevent.Timeout(1)
+            try:
+                while awaiting_child:
+                    gevent.sleep(0.01)
+                # We should now be able to waitpid() for an arbitrary child
+                wpid, status = os.waitpid(-1, os.WNOHANG)
+                if wpid != pid:
+                    raise AssertionError("Failed to wait on a child pid forked with a function",
+                                         wpid, pid, func)
+
+                # And a second call should raise ECHILD
+                try:
+                    wpid, status = os.waitpid(-1, os.WNOHANG)
+                    raise AssertionError("Should not be able to wait again")
+                except OSError as e:
+                    assert e.errno == errno.ECHILD
+            except gevent.Timeout as t:
+                if timeout is not t:
+                    raise
+                raise AssertionError("Failed to wait using", func)
+            finally:
+                timeout.cancel()
+    sys.exit(0)
 else:
     print("No SIGCHLD, not testing")